Transferred the Custom Tabs code to a separate module.
