+++ app/build.gradle
@@ -166,6 +166,8 @@ fladle {
 }
 
 dependencies {
+	implementation project(":custom-tabs-impl")
+	implementation project(":custom-tabs-api")
 	implementation project(":history-impl")
 	implementation project(":history-api")
 	implementation project(":data-store-impl")
