+++ custom-tabs/custom-tabs-api/src/main/java/com/duckduckgo/customtabs/api/CustomTabDetector.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2024 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.customtabs.api
+
+/**
+ * Interface for detecting if a tab is a custom tab.
+ */
+interface CustomTabDetector {
+    /**
+     * Checks if the current tab is a custom tab.
+     *
+     * @return true if the current tab is a custom tab, false otherwise.
+     */
+    fun isCustomTab(): Boolean
+
+    /**
+     * Sets the current tab as a custom tab or a regular tab.
+     *
+     * @param value true to set the current tab as a custom tab, false to set it as a regular tab.
+     */
+    fun setCustomTab(value: Boolean)
+}
