+++ custom-tabs/readme.md
@@ -1,0 +1,9 @@
+# Feature Name
+Implementation of the Custom Tab feature. This is used to open a custom tab from another application when the DuckDuckGo browser is 
+set as default.
+
+## Who can help you better understand this feature?
+- Ana Capatina
+
+## More information
+- [Asana: DuckDuckGo Custom Tab](https://app.asana.com/0/1200905986587319/1206880175454747/f)
